+++ packages/granularity/src/components/GrCard/config.ts
@@ -1,5 +1,8 @@
 import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
 
+import { grCardSafelist } from './safelist'
+
 export const grCardConfig = defineGranularComponent(import.meta.url, {
   name: 'GrCard',
+  safelist: grCardSafelist,
 })